refactor(elements): Sign up consolidation interim commit
