+++ packages/elements/src/internals/machines/sign-up/index.ts
@@ -1,8 +1,5 @@
 export { SignUpRouterMachine, SignUpRouterMachineId } from './router.machine';
-export { SignUpVerificationMachine, SignUpVerificationMachineId } from './verification.machine';
 
 export type { TSignUpRouterMachine } from './router.machine';
-export type { TSignUpVerificationMachine } from './verification.machine';
 
 export type * from './router.types';
-export type * from './verification.types';
