+++ packages/elements/src/internals/machines/sign-up/router.types.ts
@@ -1,5 +1,5 @@
 import type { SignUpResource } from '@clerk/types';
-import type { ActorRefFrom, ErrorActorEvent, SnapshotFrom, StateMachine } from 'xstate';
+import type { ActorRefFrom, DoneActorEvent, ErrorActorEvent, SnapshotFrom, StateMachine } from 'xstate';
 
 import type { TFormMachine } from '~/internals/machines/form';
 import type {
@@ -20,11 +20,26 @@ import type {
 
 // ---------------------------------- Tags ---------------------------------- //
 
+export type SignUpVerificationVerificationCategoryTags = 'verification:category:code' | 'verification:category:link';
+export type SignUpVerificationVerificationMethodTags = 'verification:method:email' | 'verification:method:phone';
+export type SignUpVerificationVerificationTypeTags =
+  | 'verification:email_link'
+  | 'verification:email_code'
+  | 'verification:phone_code';
+
+export type SignUpVerificationTags =
+  | SignUpVerificationVerificationCategoryTags
+  | SignUpVerificationVerificationMethodTags
+  | SignUpVerificationVerificationTypeTags;
+
+export type SignUpVerificationFriendlyTags = 'code' | 'email_link' | 'email_code' | 'phone_code';
+
 export const SignUpRouterStates = {
   attempting: 'state:attempting',
   loading: 'state:loading',
   pending: 'state:pending',
   preparing: 'state:preparing',
+  active: 'state:active',
 } as const;
 
 export const SignUpRouterSteps = {
@@ -39,6 +54,7 @@ export type SignUpRouterStates = keyof typeof SignUpRouterStates;
 export type SignUpRouterSteps = keyof typeof SignUpRouterSteps;
 
 export type SignUpRouterTags =
+  | SignUpVerificationTags
   | (typeof SignUpRouterSteps)[keyof typeof SignUpRouterSteps]
   | (typeof SignUpRouterStates)[keyof typeof SignUpRouterStates];
 
@@ -67,6 +83,7 @@ export type SignUpRouterResetStepEvent = BaseRouterResetStepEvent;
 export type SignUpRouterLoadingEvent = BaseRouterLoadingEvent<
   'start' | 'continue' | 'verifications' | 'reset-password' | 'forgot-password' | 'choose-strategy' | 'error'
 >;
+export type SignUpRouterVerificationRetryEvent = { type: 'RETRY' };
 export type SignUpRouterSubmitEvent = { type: 'SUBMIT' };
 export type SignUpRouterSetClerkEvent = BaseRouterSetClerkEvent;
 
@@ -77,6 +94,39 @@ export interface SignUpRouterInitEvent extends BaseRouterInput {
 }
 
 export type SignUpRouterNavigationEvents = SignUpRouterStartEvent | SignUpRouterPrevEvent;
+
+export type SignUpVerificationEmailLinkVerifiedEvent = { type: 'EMAIL_LINK.VERIFIED'; resource: SignUpResource };
+export type SignUpVerificationEmailLinkUnverifiedEvent = { type: 'EMAIL_LINK.UNVERIFIED'; resource: SignUpResource };
+export type SignUpVerificationEmailLinkExpiredEvent = { type: 'EMAIL_LINK.EXPIRED'; resource: SignUpResource };
+export type SignUpVerificationEmailLinkTransferrableEvent = {
+  type: 'EMAIL_LINK.TRANSFERRABLE';
+  resource: SignUpResource;
+};
+export type SignUpVerificationEmailLinkRestartEvent = { type: 'EMAIL_LINK.RESTART' };
+export type SignUpVerificationEmailLinkFailedEvent = {
+  type: 'EMAIL_LINK.FAILED';
+  resource: SignUpResource;
+  error: Error;
+};
+
+export type SignUpVerificationSubmitEvent = { type: 'SUBMIT'; action: 'submit' };
+export type SignUpVerificationNextEvent = { type: 'NEXT'; resource?: SignUpResource };
+export type SignUpVerificationRetryEvent = { type: 'RETRY' };
+
+export type SignUpVerificationEmailLinkEvent =
+  | SignUpVerificationEmailLinkVerifiedEvent
+  | SignUpVerificationEmailLinkUnverifiedEvent
+  | SignUpVerificationEmailLinkExpiredEvent
+  | SignUpVerificationEmailLinkRestartEvent
+  | SignUpVerificationEmailLinkFailedEvent;
+
+export type SignUpVerificationEvents =
+  | DoneActorEvent
+  | ErrorActorEvent
+  | SignUpVerificationRetryEvent
+  | SignUpVerificationSubmitEvent
+  | SignUpVerificationNextEvent
+  | SignUpVerificationEmailLinkEvent;
 
 export type SignUpRouterEvents =
   | ErrorActorEvent
@@ -91,11 +141,20 @@ export type SignUpRouterEvents =
   | SignUpRouterResetStepEvent
   | SignUpRouterLoadingEvent
   | SignUpRouterSetClerkEvent
-  | SignUpRouterSubmitEvent;
+  | SignUpRouterVerificationRetryEvent
+  | SignUpRouterSubmitEvent
+  | SignUpVerificationEmailLinkVerifiedEvent
+  | SignUpVerificationEmailLinkUnverifiedEvent
+  | SignUpVerificationEmailLinkExpiredEvent
+  | SignUpVerificationEmailLinkTransferrableEvent
+  | SignUpVerificationEmailLinkRestartEvent
+  | SignUpVerificationEmailLinkFailedEvent;
 
 // ---------------------------------- Delays ---------------------------------- //
 
 export const SignUpRouterDelays = {
+  emailLinkTimeout: 300_000, // 5 minutes
+  resendableTimeout: 1_000, // 1 second
   polling: 300_000, // 5 minutes
 } as const;
 
@@ -108,6 +167,9 @@ export type SignUpRouterLoadingContext = Omit<SignUpRouterLoadingEvent, 'type'>;
 export interface SignUpRouterContext extends BaseRouterContext {
   formRef: ActorRefFrom<TFormMachine>;
   loading: SignUpRouterLoadingContext;
+  resendable: boolean;
+  resendableAfter: number;
+  resource: SignUpResource;
   signInPath: string;
   ticket: string | undefined;
 }
